+++ testdata/collide2/alpha/alpha.go
@@ -1,0 +1,5 @@
+package alpha
+
+type Config struct{ Host string }
+
+func Alpha() Config { return Config{Host: "a"} }