+++ testdata/collide2/beta/beta.go
@@ -1,0 +1,5 @@
+package beta
+
+type Config struct{ Port int }
+
+func Beta() Config { return Config{Port: 1} }